backend_rest/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.1
django-filter==2.2.0
django-multiselectfield==0.1.11
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
oauthlib==3.1.0
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.7
